Truncate column examples to save on session space
The plugin has historically told users to alter the Omeka session table
definition to allow more storage. This is only necessary because the
plugin stores the entire content of each "column" in the session before
showing the mapping page. Only the first few characters are ever
visible, so we can just truncate the column example texts before
returning them and the length of the transcript will no longer matter.